requirements/prod.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
pillow==10.3.0
reportlab==4.2.0
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.6.0
